Initial commit.
The swampbackend is almost done. Which is to say that the end is in
sight, and I can see how to do the few simple things that are still
missing.

Of course, there's a lot of lockdown still to do, security etc.

But it's time to shoot the engineers, so to speak, and get started on
the angular app that is supposed to go with this monstrosity.