Minor change: call super with ordered data instead.
This is, I think, the absolute minimum change to make the sub-list
ordering conform to what I want. I could even generalize it, but that
would require overriding the "__init__" function to take a
"order_by='word__word'" style argument, saving it as an object property,
and then refering to it in the to_representation call. Worth doing if I
have to do this again though.